Slightly rework Rakefile.
'build' now depends on 'test' and only builds the gem, doesn't install it.
Dependency on 'test' means gem won't be built if tests don't pass. Additionally,
'default' now calls out to both 'build' (which in turn calls out to 'test') and
'doc' so that running rake equals to a "complete build".

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>